réorganise le code et ajoute une date de fin pour la sélection des logs
