+++ pkg/stats/resources/sql/create_tables_and_views.sql
