+++ pkg/stats/resources/tests/token
@@ -1,0 +1,1 @@
+eyJhbGciOiJSUzI1NiIsInR5cCIgOiAiSldUIiwia2lkIiA6ICJtb3hEcWt2QzRVMy1LVUVLNmJoanJ0SzdPRWNlbzlqejhnNERnS0JQeWo0In0.eyJleHAiOjE2ODQ1NzYxOTQsImlhdCI6MTY4NDU3NTI5NCwiYXV0aF90aW1lIjoxNjg0NTc1MjkzLCJqdGkiOiIzNjJmYmM0Ni0wYjczLTRmZWItOWM0ZS0wYzczNzA3MjJlMmQiLCJpc3MiOiJodHRwczovL3Rlc3Rpbmcuc2lnbmF1eC1mYWlibGVzLmRldi9hdXRoL3JlYWxtcy9tYXN0ZXIiLCJhdWQiOiJhY2NvdW50Iiwic3ViIjoiODljYWIzMDgtMmM1YS00MDRmLTlhZDktM2E1MWM3MWNjZGE4IiwidHlwIjoiQmVhcmVyIiwiYXpwIjoic2lnbmF1eGZhaWJsZXMiLCJub25jZSI6ImNhN2I4MTY4LTM2NWEtNDZkMC04ZGFjLTZjMzE5OGFiNWM0MyIsInNlc3Npb25fc3RhdGUiOiIyYjEyYTgzMi1kOWE5LTQ3MjYtYTc0Yi04NDk4Nzc1OWZmZmEiLCJhY3IiOiIxIiwiYWxsb3dlZC1vcmlnaW5zIjpbImh0dHA6Ly9sb2NhbGhvc3Q6ODA4MCIsImh0dHBzOi8vdGVzdGluZy5zaWduYXV4LWZhaWJsZXMuZGV2Il0sInJlYWxtX2FjY2VzcyI6eyJyb2xlcyI6WyJkZWZhdWx0LXJvbGVzLW1hc3RlciIsIm9mZmxpbmVfYWNjZXNzIiwidW1hX2F1dGhvcml6YXRpb24iXX0sInJlc291cmNlX2FjY2VzcyI6eyJzaWduYXV4ZmFpYmxlcyI6eyJyb2xlcyI6WyI4OCIsIjg5IiwiMDEiLCIwMiIsIjAzIiwiMDQiLCIwNSIsIjA2IiwiMDciLCIwOCIsIjA5Iiwic2NvcmUiLCI5MCIsIjkxIiwiOTIiLCI5MyIsIjk0IiwiOTUiLCIxMCIsIjExIiwiMTIiLCIxMyIsIjE0IiwiMTUiLCIxNiIsIjE3IiwiMTgiLCIxOSIsImRnZWZwIiwiMjEiLCIyMiIsIjIzIiwiMjQiLCIyNSIsIjI2IiwiMjciLCIyOCIsIjI5IiwicGdlIiwiMkEiLCIyQiIsIjMwIiwiMzEiLCIzMiIsIjMzIiwiMzQiLCJkZXRlY3Rpb24iLCIzNSIsIjM2IiwiMzciLCJiZGYiLCIzOCIsIjM5IiwiNDAiLCI0MSIsIjQyIiwiNDMiLCI0NCIsInVyc3NhZiIsIjQ1IiwiNDYiLCI0NyIsIjQ4IiwiNDkiLCI1MCIsIjUxIiwiNTIiLCI1MyIsIjU0IiwiNTUiLCI1NiIsIjU3IiwiNTgiLCI1OSIsIkZyYW5jZSBlbnRpw6hyZSIsIjYwIiwiNjEiLCI2MiIsIjYzIiwiNjQiLCI2NSIsIjY2IiwiNjciLCI2OCIsIjY5IiwiOTcxIiwiOTcyIiwiOTczIiwiOTc0IiwiOTc2IiwiNzAiLCI3MSIsIjcyIiwiNzMiLCI3NCIsIjc1IiwiNzYiLCI3NyIsIjc4IiwiNzkiLCJ3ZWthbiIsIjgwIiwiODEiLCI4MiIsIjgzIiwiODQiLCI4NSIsIjg2IiwiODciXX0sImFjY291bnQiOnsicm9sZXMiOlsibWFuYWdlLWFjY291bnQiLCJtYW5hZ2UtYWNjb3VudC1saW5rcyIsInZpZXctcHJvZmlsZSJdfX0sInNjb3BlIjoib3BlbmlkIGVtYWlsIHByb2ZpbGUiLCJzaWQiOiIyYjEyYTgzMi1kOWE5LTQ3MjYtYTc0Yi04NDk4Nzc1OWZmZmEiLCJlbWFpbF92ZXJpZmllZCI6dHJ1ZSwibmFtZSI6IkNocmlzdG9waGUgTklOVUNDSSIsInByZWZlcnJlZF91c2VybmFtZSI6ImNocmlzdG9waGUubmludWNjaUBiZXRhLmdvdXYuZnIiLCJnaXZlbl9uYW1lIjoiQ2hyaXN0b3BoZSIsImZhbWlseV9uYW1lIjoiTklOVUNDSSIsImVtYWlsIjoiY2hyaXN0b3BoZS5uaW51Y2NpQGJldGEuZ291di5mciJ9.Ynbet0C3jUOv-Sy4ul0Rbgrt-qCx3GS0E89E-vA-NqMovficTwavDzybxh_RfjsG2omNSeuueLJwVU5nplwO-ohm09o7TveP1pehy1GeFG1LYXP2MK68OaD9WQo5DfUtrFUtRsvy_K9aEV_cM24MjN689tGeo-FerLobZNUj0cmnsLrfuZwLe2V7TxZIzaEkX2i96jzlEbtuGfCOimTTiw5lpEkiJPBaJFm1BsvmMdogMKQeBgJA9Fm04__gSf53RopVqIP2ZtfjuiuvxA2-83ZM0h1iej150v6RH8PwZX3RGsvBtCzXrYPCy_4L0oU5Jl2-tyApoUveI5bMuKIr2w
